refactor: move file to components folder
