+++ components/common/inventory-filters.tsx
